feat: add initial post search autocomplete
This autocomplete feature sets up an endpoint to take a post search
query, parses it and return a set of AutocompleteItem(s) populated
with matching tags and other search functions. Only a small subset
of search functions with there validators have been added, they
still need to be tested and more search function will be added at
a later time.

The new endpoint at posts/autocomplete/ returns a simple list of
these items in HTML to be used with HTMX for dynamic autocompletion
during post searching.

Note that tag categories were also refactored into a simple Enum
instead of operating as a distinct table in the database. It may
be slightly less flexible but is easier to work with and will
probably be more maintainable this way.