refactor: replace Ratings tags with Enum
An Enum makes much more semantic sense for Post Ratings. Applying them
in tags is a complication because tags are intended to be added to posts
ad hoc. Any tag should be applicable to any post, but a Rating should
only apply one of a set of choices. The Tags still retain there optional
rating_level attribute which can be used for setting some minimum rating
level by tag, but Ratings are now handled directly in Posts as an integer
for simplicity.