feat: add initial tagset builder component
This component will be the basis for tag selection in the Post creation
workflow on the Upload page, but also for use on Post edit pages.

Also adds corresponding Form for parameter validation. Note that sending
the list of tags, requires some manual parsing of the POST parameters
to get a list of Tag IDs, otherwise Django just uses the last ID from
the list of Tag ID <input>s.