refactor: lift some filtering logic out of views into models
- Following the guidance of this article https://spookylukey.github.io/django-views-the-right-way/thin-views.html
which recommend "thin views", i.e. "business" logic should be lifting out
of views into the model layer. This creates a sort of semantic API for
accessing data from the DB consistently and will be helpful when implementing
the API to keep querying consistent.
- Also adds a baseline decorators.py module
- Also fixes Post filter results. Before, it was returning Post with any
of the provided tags similar to combining with OR. Now results are only
shown for Post tagged with _all_ the provided tags.