fix: delete button fails on global collections page
- Refactor collection thumbnail to use a slot for the colleciton delete
button. This makes the collection deletion function only available on
the user's collection page which I think makes more sense. Users's don't
need to be able to delete collection from the global view, it just
clutters up the UI unnecessarily.
- Also restyles the colleciton thumbnail, putting the buttons on the
bottom and forcing the visibility to the top-right corner with absolute
positioning.